[IceTV] Add German genres & server selection
Select the IceTV server from a list of services in the setup wizard.

Save server name in config.plugins.icetv.server.name instead of
local API variable.

Add IceTV German genre eit_id remap table (for icetv.de), and switch
between remap tables on the value of config.plugins.icetv.member.country.

Encode category/genre names in UTF-8 (previously unnecessary for
icetv.com.au data.

Make the fetch of show["category"]["eit"] more robust - set it to
"0" if the "eit" attribute is missing.

Switch between QR code pixmaps based on
config.plugins.icetv.member.country.

Re-enable printing of debug log error message when there's a failure
to remap a genre id, but only show the first failure for each genre
name.

Use new country code field in region table to set
config.plugins.icetv.member.country instead of using the country
name, which has issues with translation.

Default config.plugins.icetv.member.country to "AUS" for
backward-compatibility for existing Australian IceTV users.

Set config.plugins.icetv.member.country from the user's region
in IceTVLogin so that the country code is set whan the user chooses
Existing or trial user in the setup. This is over-ridden by a do-nothing
method in IceTVCreateLogin because in that case, the country code will
have already been set in IceTVNewUserSetup.

Change titlebartext= to windowTitle= in the initialisation of ChoiceBox in
IceTVMain for improved compatibility between OpenViX and OpenATV, and
add a fallback initialisation for Beyonwiz, which doesn't
support either.

Set skinName in IceTVOldUserSetup & IceTVCreateLogin and  to their
respective superclasses so that they can use the same skins as the
superclasses if they have a screen skin in the current selected skin.

In the IceTV setup wizard, save config variables only on completion
of the wizard, and reset the variables when the screens that set
them are cancelled.

Bump version string.

[Components.Converter.genre]

Add an IceTV genre table for icetv.de, and generalise switching
between IceTV  genre tables based on country.

In __remapCountry() use the IceTV country code if IceTV is in use and
no country code is specified. This fixes an exception in __remapCountry()
when used with some enigma2 apps.

In getGenreStringLong() pass the country code to getGenreStringSub().