[IceTV] Name lookup fallback for channel triples
At Daniel Hall@IceTV's request,

If the IceTV channel table's "name" or "known_names" fields contain
a channel name that's in the enigma2 channel DB, but the enigma2
channel DB has triples for that name that aren't in the IceTV
channel, add those triples to the IceTV channel table.

This allows IceTV to use this channel name fallback to fill in
triples that are not (yet) in the IceTV database.